requirements.txt
gradio>=4.0.0
requests>=2.31.0
pydantic>=2.6.0
python-dotenv>=1.0.1
pillow>=10.0.0
